Validation working
Added the common1900.xsd schema. This is a modified version of the schema from the MNX clone repository.
Added validation code to MNX_Main_Form1.cs
Added __1_helloHTML.mnx, containing correct <mnx> attributes. This file contains a <text-block> element that is read into a TextBlock constructor by the Directions constructor, but not yet in the schema.